+++ src/GameMain.java
@@ -1,0 +1,7 @@
+
+public class GameMain {
+	
+	public static void main(String[] args) {
+		GameGUI gui = new GameGUI("Numbers Game");
+	}
+}
